+++ lib/asan/tests/asan_test_main.cc
@@ -19,8 +19,7 @@ extern "C" const char* __asan_default_options() {
 #if SANITIZER_MAC
   // On Darwin, we default to `abort_on_error=1`, which would make tests run
   // much slower. Let's override this and run lit tests with 'abort_on_error=0'.
-  // Also, make sure we do not overwhelm the syslog while testing.
-  return "symbolize=false:abort_on_error=0:log_to_syslog=0";
+  return "symbolize=false:abort_on_error=0";
 #else
   return "symbolize=false";
 #endif
