+++ lib/sanitizer_common/sanitizer_common.h
@@ -162,7 +162,6 @@ void SetLowLevelAllocateCallback(LowLevelAllocateCallback callback);
 // IO
 void RawWrite(const char *buffer);
 bool ColorizeReports();
-void RemoveANSIEscapeSequencesFromString(char *buffer);
 void Printf(const char *format, ...);
 void Report(const char *format, ...);
 void SetPrintfAndReportCallback(void (*callback)(const char *));
@@ -649,27 +648,13 @@ enum AndroidApiLevel {
   ANDROID_POST_LOLLIPOP = 23
 };
 
-void WriteToSyslog(const char *buffer);
-
-#if SANITIZER_MAC
-void LogFullErrorReport(const char *error_message_buffer);
-#else
-INLINE void LogFullErrorReport(const char *error_message_buffer) {}
-#endif
-
-#if SANITIZER_LINUX || SANITIZER_MAC
-void WriteOneLineToSyslog(const char *s);
-#else
-INLINE void WriteOneLineToSyslog(const char *s) {}
-#endif
-
 #if SANITIZER_LINUX
 // Initialize Android logging. Any writes before this are silently lost.
 void AndroidLogInit();
-bool ShouldLogAfterPrintf();
+void WriteToSyslog(const char *buffer);
 #else
 INLINE void AndroidLogInit() {}
-INLINE bool ShouldLogAfterPrintf() { return false; }
+INLINE void WriteToSyslog(const char *buffer) {}
 #endif
 
 #if SANITIZER_ANDROID
